+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Changed no license field from error to warning
 - Improved error message when `cargo metadata` fails
 
 ## 0.1.0 (2021-12-08)